Updated NotificationMessages.Offer and NotificationMessages.OfferElement
Based on documentation aviable at the following link
https://spapi.cyou/en/use-case/notifications-use-case-guide-v1.html

The OfferElement class used in the ANY_OFFER_CHANGED notification have almost all properties in common with the Offer class used in B2B_ANY_OFFER_CHANGED notification.

I've reordered the properties in the same order that is on the documentation, added the missing ConditionNotes property and removed the non existent ShipsInternationally property.

After that, i've moved all the common properties into Offer class and make OfferElement a descendant of the Offer class

In this way we can create reusable code that can work for both the ANY_OFFER_CHANGED and B2B_ANY_OFFER_CHANGED classes