Replaced duplicated class NumberOfBuyBoxEligibleElements
This class was exactly the same of BuyBoxEligibleOffer. Both where present in Summary class, since the B2C feed has NumberOfBuyBoxEligibleElements property, while B2B feed has BuyBoxEligibleElements.

Those 2 properties in the 2 different feeds are the same thing, just with different names.

I've changed Summary class so that one of them (BuyBoxEligibleElements that is a "normal lenght" name) is the default, the other property (marked as obsolete to avoid it's use) is simply pointing back to the main property, allowing newtonsoft.json library to deserialize properly both feeds

I've tested this solution with real amazon data on an SQS queue where there are both B2C and B2B notifies and it works as expected